Put definitions beside their declarations, or declarations beside their definitions
Acts on the eight findings the decl/def audit reported. Six of them could not be fixed the
way the instruction read, and the reason is the useful part.

Only two definitions could move DOWN. dt_iop_buffer_dsc_update_bpp() and
dt_iop_buffer_dsc_to_bpp() touch dt_iop_buffer_dsc_t and nothing else, so they left
develop/format.c for a new pixel/format.c, beside the header that declares them.

The other six operate on develop/-owned types -- dt_iop_module_t, dt_dev_pixelpipe_t,
dt_dev_history_item_t -- so moving their definitions down would have dragged layer 5 into
layers 1 and 2. Their DECLARATIONS moved up instead, which puts declaration and definition
together just the same:

  default_input_format, default_output_format, default_blend_colorspace
      pixel/format.h -> develop/imageop.h; all 25 callers are iop/, which includes it already
  dt_history_duplicate
      common/history.h -> develop/dev_history.h; its only common/ caller, history_merge.c,
      already includes that header

The payoff is what pixel/format.h no longer has to say: the three forward declarations of
dt_dev_pixelpipe_t, dt_dev_pixelpipe_iop_t and dt_iop_module_t existed solely so a layer-2
header could declare functions over layer-5 types. Verified they had no other use by
grepping outside comments -- the first check matched the comment I had just written.

The remaining two are correct as they stand and are documented rather than moved:
common/pixelpipe_cache_alloc.h declares two functions defined in develop/pixelpipe_cache.c,
but the allocation macros built on them are used from pixel/ (14 files), common/ (6), math/
and imageio/. Moving the declarations up forces all those layers to include develop/; moving
the definitions down drags cache internals into layer 1. An opaque handle passed through
costs neither, which is what it already is.
