+++ src/CMakeLists.txt
@@ -85,13 +85,13 @@ FILE(GLOB SOURCE_FILES
   "pixel/guided_filter.c"
   "common/history.c"
   "common/history_actions.c"
-  "common/history_merge.c"
+  "develop/history_merge.c"
   "gui/common/collection_gui.c"
   "gui/common/database_gui.c"
   "gui/common/film_gui.c"
   "gui/common/folder_survey_gui.c"
   "gui/common/history_actions_gui.c"
-  "gui/common/history_merge_gui.c"
+  "gui/develop/history_merge_gui.c"
   "gui/common/styles_gui.c"
   "common/history_snapshot.c"
   "common/topological_sort.c"
@@ -149,8 +149,9 @@ FILE(GLOB SOURCE_FILES
   "common/histogram.c"
   "common/undo.c"
   "common/usermanual_url.c"
-  "common/iop_order.c"
-  "common/iop_profile.c"
+  "develop/iop_order.c"
+  "develop/iop_profile.c"
+  "pixel/iop_profile.c"
   "common/dng_opcode.c"
   "common/conf.c"
   "control/control.c"
@@ -197,6 +198,7 @@ FILE(GLOB SOURCE_FILES
   "develop/masks/masks_history.c"
   "develop/masks/polygon.c"
   "develop/format.c"
+  "pixel/format.c"
   "gui/dtgtk/preview_window.c"
   "gui/dtgtk/thumbnail.c"
   "gui/dtgtk/thumbtable_info.c"
