+++ src/develop/dev_history.h
@@ -18,7 +18,7 @@
 */
 #include "system/atomic.h"
 #include "common/history.h"
-#include "common/history_merge.h"
+#include "develop/history_merge.h"
 
 #include <glib.h>
 
@@ -659,4 +659,10 @@ gboolean dt_dev_transient_params_active(struct dt_develop_t *dev, const struct d
 #ifdef __cplusplus
 }
 #endif
+
+/** Deep-copy a history list. Defined in dev_history.c; declared in common/history.h until it
+ *  was moved here, since it walks dt_dev_history_item_t and resolves modules through
+ *  dt_iop_get_module() -- neither of which layer 1 can see. */
+GList *dt_history_duplicate(GList *hist);
+
 #endif // DT_DEVELOP_DEV_HISTORY_H
