+++ src/gui/develop/history_merge_gui.h
